slideware: add overview slide set
Add overview slide set, intended to give a breif overview of what the Linux
kernel is, how it acts as a component in the system, how the source code is
organized and how to do a build.

Also remove --to-latex option when converting svgs to pdfs, as somewhow
the text gets completely misformatted when including it via inkscape generated
.tex files.

Including pdfs instead and converting text to paths during svg-pdf
conversion solves the problem.

Signed-off-by: Hans Holmberg <hans@pixelmunchies.com>